feat(glob): split vite_glob into env/path modules; env globs support negation
`vite_glob` wrapped `wax`, a filesystem path-glob engine, but most callers
used it to match environment-variable *names* — flat strings, not paths —
by pushing `&str` through `is_match(impl AsRef<Path>)`. That is semantically
muddy, and `wax` exposes no option to disable its path semantics.

Reorganize the crate into two modules, each with its own matchers and error
type:

- `vite_glob::env` — `EnvGlob` (one literal pattern) and `EnvGlobSet` (a set,
  backed by `globset`) with `literal_separator(false)` and
  `case_insensitive(cfg!(windows))`, so `*`/`?`/`[...]`/`{a,b}` are pure
  flat-string wildcards with the platform's env case rules. `EnvGlobSet`
  supports **negation**: a `!`-prefixed pattern excludes, and a name matches
  when it matches an include and no exclude (e.g. `["VITE_*", "!VITE_SECRET"]`).
  Errors: `EnvGlobError`.
- `vite_glob::path` — `PathGlobSet` (the former `GlobPatternSet`, renamed),
  still `wax`-backed with unchanged gitignore (first/last-match-wins)
  semantics. Errors: `PathGlobError`.

This mirrors how Turborepo splits env (regex) from path (wax) matching, and how
Turbo treats `!` in env wildcards. Non-negated env results are unchanged —
globset matches the same `*`/`?`/`[...]`/`{}` syntax wax did for separator-free
names, including SENSITIVE_PATTERNS and the Unix/Windows case split (covered by
existing envs.rs tests).

Migrate the pre-existing callers: `envs.rs` uses `EnvGlobSet` for its
multi-pattern env matching — so `env: ["VITE_*", "!VITE_SECRET"]` in a task
config now excludes `VITE_SECRET` (previously `!` patterns were dropped with a
warning) — and `vite_workspace` switches to `PathGlobSet` / `PathGlobError`.
The runner-aware env-matching sites consume the single-pattern `EnvGlob` in the
stacked PR.
