+++ crates/vite_glob/src/path.rs
@@ -1,0 +1,375 @@
+//! Glob matching for filesystem **paths** with gitignore semantics.
+//!
+//! Backed by `wax`. `!`-prefixed patterns negate. With no negation present,
+//! first-match-wins; with any negation, last-match-wins (gitignore).
+
+#[expect(clippy::disallowed_types, reason = "wax::Glob::is_match requires std::path::Path")]
+use std::path::Path;
+
+use wax::{Glob, Program};
+
+/// Error compiling a path glob pattern.
+#[derive(Debug, thiserror::Error)]
+#[error(transparent)]
+pub struct PathGlobError(#[from] wax::BuildError);
+
+/// Matches filesystem paths against an ordered set of glob patterns.
+///
+/// If there are no negated patterns, it will follow the first match wins semantics.
+/// Otherwise, it will follow the last match wins semantics.
+#[derive(Debug)]
+pub struct PathGlobSet<'a> {
+    /// (`glob_pattern`, `match_or_not`)
+    patterns: Vec<(Glob<'a>, bool)>,
+    has_negated: bool,
+}
+
+impl<'a> PathGlobSet<'a> {
+    /// # Errors
+    /// Returns an error if any glob pattern is invalid.
+    pub fn new<I, S>(match_patterns: I) -> Result<Self, PathGlobError>
+    where
+        I: IntoIterator<Item = &'a S>,
+        S: AsRef<str> + 'a + ?Sized,
+    {
+        let mut patterns = Vec::new();
+        let mut has_negated = false;
+        for pattern in match_patterns {
+            let pattern_str = pattern.as_ref();
+            if let Some(negated) = pattern_str.strip_prefix('!') {
+                // negated pattern, ignore the path
+                patterns.push((Glob::new(negated)?, false));
+                // set to true to follow last match wins semantics
+                has_negated = true;
+            } else {
+                // positive pattern, match the path
+                patterns.push((Glob::new(pattern_str)?, true));
+            }
+        }
+        Ok(Self { patterns, has_negated })
+    }
+
+    #[expect(clippy::disallowed_types, reason = "wax::Glob::is_match requires std::path::Path")]
+    pub fn is_match(&self, path: impl AsRef<Path>) -> bool {
+        let mut should_match = false; // Default: don't match
+        for (glob, match_or_not) in &self.patterns {
+            if glob.is_match(path.as_ref()) {
+                should_match = *match_or_not;
+                if !self.has_negated {
+                    // first match wins semantics
+                    break;
+                }
+            }
+        }
+        should_match
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_match_ignores_node_modules() -> Result<(), PathGlobError> {
+        let patterns = vec![
+            // ignore all paths
+            "**/*",
+            // keep node_modules directories themselves
+            "!**/node_modules",
+            "!node_modules",
+            // keep lock files and package.json
+            "!**/package.json",
+            "!**/package-lock.json",
+            "!**/yarn.lock",
+            "!**/pnpm-lock.yaml",
+        ];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should ignore paths inside node_modules
+        assert!(ignores.is_match("node_modules/react/index.js"));
+        assert!(ignores.is_match("apps/web/node_modules/react/index.js"));
+        assert!(ignores.is_match("packages/cli/node_modules/@types/node/index.d.ts"));
+
+        // Should ignore paths outside node_modules
+        assert!(ignores.is_match("src/index.js"));
+        assert!(ignores.is_match("tsbuildinfo.json"));
+
+        // Should NOT ignore node_modules directories themselves (due to negation)
+        assert!(!ignores.is_match("node_modules"));
+        assert!(!ignores.is_match("apps/web/node_modules"));
+        assert!(!ignores.is_match("packages/cli/node_modules"));
+
+        // Should NOT ignore lock files and package.json
+        assert!(!ignores.is_match("package.json"));
+        assert!(!ignores.is_match("apps/web/package.json"));
+        assert!(!ignores.is_match("package-lock.json"));
+        assert!(!ignores.is_match("apps/web/yarn.lock"));
+        assert!(!ignores.is_match("pnpm-lock.yaml"));
+        assert!(!ignores.is_match("node_modules/react/package.json"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_with_file_patterns() -> Result<(), PathGlobError> {
+        let patterns = vec!["*.log", "**/*.tmp", "!important.log"];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should ignore matching files
+        assert!(ignores.is_match("debug.log"));
+        assert!(ignores.is_match("error.log"));
+        assert!(ignores.is_match("temp/file.tmp"));
+        assert!(ignores.is_match("deep/nested/path/cache.tmp"));
+        #[expect(clippy::disallowed_types, reason = "wax::Glob::is_match requires std::path::Path")]
+        {
+            assert!(ignores.is_match(String::from("deep/nested/path/cache.tmp")));
+            assert!(ignores.is_match(Path::new("deep/nested/path/cache.tmp")));
+        }
+
+        // Should NOT ignore negated patterns
+        assert!(!ignores.is_match("important.log"));
+
+        // Should NOT ignore non-matching files
+        assert!(!ignores.is_match("file.txt"));
+        assert!(!ignores.is_match("logs/file.txt"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_directory_patterns() -> Result<(), PathGlobError> {
+        let patterns = vec!["dist/**", "build/**", "!dist/public/**"];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should ignore paths in dist and build
+        assert!(ignores.is_match("dist/bundle.js"));
+        assert!(ignores.is_match("dist/assets/style.css"));
+        assert!(ignores.is_match("build/output.js"));
+        assert!(ignores.is_match("build/assets/image.png"));
+
+        // Should NOT ignore negated paths
+        assert!(!ignores.is_match("dist/public/index.html"));
+        assert!(!ignores.is_match("dist/public/assets/logo.png"));
+
+        // Should NOT ignore paths outside target directories
+        assert!(!ignores.is_match("src/index.js"));
+        assert!(!ignores.is_match("public/index.html"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_complex_patterns() -> Result<(), PathGlobError> {
+        let patterns = vec![
+            "**/*.test.js",
+            "**/*.spec.ts",
+            "**/test/**",
+            "**/tests/**",
+            "!**/integration/tests/**",
+        ];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should ignore test files
+        assert!(ignores.is_match("src/utils.test.js"));
+        assert!(ignores.is_match("components/Button.spec.ts"));
+        assert!(ignores.is_match("lib/test/helper.js"));
+        assert!(ignores.is_match("src/tests/unit/math.js"));
+
+        // Should NOT ignore negated patterns
+        assert!(!ignores.is_match("integration/tests/e2e.js"));
+        assert!(!ignores.is_match("integration/tests/api/user.js"));
+
+        // Should NOT ignore non-test files
+        assert!(!ignores.is_match("src/index.js"));
+        assert!(!ignores.is_match("lib/utils.js"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_empty_patterns() -> Result<(), PathGlobError> {
+        let patterns: Vec<&str> = vec![];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should not ignore anything with empty patterns
+        assert!(!ignores.is_match("node_modules/package.json"));
+        assert!(!ignores.is_match("src/index.js"));
+        assert!(!ignores.is_match("dist/bundle.js"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_with_wildcards() -> Result<(), PathGlobError> {
+        let patterns = vec!["*.{js,ts,jsx,tsx}", "!index.js", "!main.ts"];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should ignore matching extensions
+        assert!(ignores.is_match("utils.js"));
+        assert!(ignores.is_match("component.tsx"));
+        assert!(ignores.is_match("service.ts"));
+        assert!(ignores.is_match("App.jsx"));
+
+        // Should NOT ignore negated files
+        assert!(!ignores.is_match("index.js"));
+        assert!(!ignores.is_match("main.ts"));
+
+        // Should NOT ignore other extensions
+        assert!(!ignores.is_match("styles.css"));
+        assert!(!ignores.is_match("data.json"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_dotfiles() -> Result<(), PathGlobError> {
+        let patterns = vec![".*", "!.gitignore", "!.env.example"];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Should ignore dotfiles
+        assert!(ignores.is_match(".env"));
+        assert!(ignores.is_match(".DS_Store"));
+        assert!(ignores.is_match(".vscode"));
+
+        // Should NOT ignore negated dotfiles
+        assert!(!ignores.is_match(".gitignore"));
+        assert!(!ignores.is_match(".env.example"));
+
+        // Should NOT ignore regular files
+        assert!(!ignores.is_match("README.md"));
+        assert!(!ignores.is_match("src/index.js"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_root_patterns() -> Result<(), PathGlobError> {
+        // Note: wax doesn't support leading / for root patterns like gitignore
+        // Using glob patterns that work with wax
+        let patterns = vec![
+            "**/dist", // Match dist at any level
+            "!dist/public",
+            "**/node_modules",
+        ];
+        let ignores = PathGlobSet::new(&patterns)?;
+        // Patterns match at any level
+        assert!(ignores.is_match("dist"));
+        assert!(ignores.is_match("src/dist")); // Also matches nested
+
+        // Negation works
+        assert!(!ignores.is_match("dist/public"));
+
+        // Node_modules patterns
+        assert!(ignores.is_match("node_modules"));
+        assert!(ignores.is_match("src/node_modules"));
+        assert!(ignores.is_match("packages/app/node_modules"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_directory_only_patterns() -> Result<(), PathGlobError> {
+        let patterns = vec![
+            "build/**",       // Match everything under build
+            "!build/keep/**", // But not under build/keep
+        ];
+        let ignores = PathGlobSet::new(&patterns)?;
+        // Directory patterns
+        assert!(ignores.is_match("build/output.js"));
+        assert!(ignores.is_match("build/assets/style.css"));
+
+        // Negated directory
+        assert!(!ignores.is_match("build/keep/important.txt"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_mixed_patterns() -> Result<(), PathGlobError> {
+        let patterns = vec![
+            "**/*.log", // Match .log files at any depth
+            "**/temp/**",
+            "node_modules/**",
+            "!**/temp/keep/**",
+            "!debug.log",
+        ];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        // Test various patterns together
+        assert!(ignores.is_match("error.log"));
+        assert!(ignores.is_match("src/app.log"));
+        assert!(!ignores.is_match("debug.log")); // Negated
+
+        assert!(ignores.is_match("temp/file.txt"));
+        assert!(ignores.is_match("src/temp/cache.dat"));
+        assert!(!ignores.is_match("temp/keep/important.txt")); // Negated
+
+        assert!(ignores.is_match("node_modules/react/index.js"));
+        assert!(ignores.is_match("node_modules/@types/node/index.d.ts"));
+
+        assert!(!ignores.is_match("src/index.js"));
+        assert!(!ignores.is_match("package.json"));
+
+        Ok(())
+    }
+
+    #[expect(
+        clippy::disallowed_types,
+        reason = "tests that is_match accepts various argument types"
+    )]
+    #[test]
+    fn test_generic_api_with_different_types() -> Result<(), PathGlobError> {
+        use vite_str::Str;
+
+        // Test with Vec<&str>
+        let patterns_str = vec!["*.log", "!important.log"];
+        let ignores_str = PathGlobSet::new(&patterns_str)?;
+        assert!(ignores_str.is_match("debug.log"));
+        assert!(!ignores_str.is_match("important.log"));
+
+        // Test with Vec<String>
+        let patterns_string = vec![String::from("*.tmp"), String::from("!keep.tmp")];
+        let ignores_string = PathGlobSet::new(&patterns_string)?;
+        assert!(ignores_string.is_match("temp.tmp"));
+        assert!(!ignores_string.is_match("keep.tmp"));
+
+        // Test with Vec<Str>
+        let patterns_vite_str = vec![Str::from("*.rs"), Str::from("!main.rs")];
+        let ignores_vite_str = PathGlobSet::new(&patterns_vite_str)?;
+        assert!(ignores_vite_str.is_match("lib.rs"));
+        assert!(!ignores_vite_str.is_match("main.rs"));
+
+        // Test with array
+        let patterns_array = ["build/**", "!build/dist/**"];
+        let ignores_array = PathGlobSet::new(&patterns_array)?;
+        assert!(ignores_array.is_match("build/src/main.js"));
+        assert!(!ignores_array.is_match("build/dist/bundle.js"));
+
+        // Test with iterator
+        let patterns_iter = ["*.md", "!README.md"].iter();
+        let ignores_iter = PathGlobSet::new(patterns_iter)?;
+        assert!(ignores_iter.is_match("CHANGELOG.md"));
+        assert!(!ignores_iter.is_match("README.md"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_match_ignores_last_matching_pattern() -> Result<(), PathGlobError> {
+        // Test that the last matching pattern wins (gitignore semantics)
+        let patterns = vec![
+            "logs/**",             // First: ignore everything in logs/
+            "!logs/important.log", // Second: don't ignore important.log
+            "logs/important.log",  // Third: ignore important.log again (this wins)
+        ];
+        let ignores = PathGlobSet::new(&patterns)?;
+
+        assert!(ignores.is_match("logs/error.log"));
+        assert!(ignores.is_match("logs/src/app.log"));
+        assert!(ignores.is_match("logs/debug.log"));
+        // The last pattern "logs/important.log" (positive) wins over "!logs/important.log" (negative)
+        assert!(ignores.is_match("logs/important.log")); // Should be ignored!
+
+        Ok(())
+    }
+}
