+++ crates/vite_task_bin/tests/e2e_snapshots/fixtures/env_negation/package.json
@@ -1,0 +1,1 @@
+{}
